Made the puzzle results grid reusable by putting it in common removing the duplications
