+++ screens/common/src/main/java/com/paulcraciunas/screens/common/model/PuzzleResult.kt
@@ -1,0 +1,7 @@
+package com.paulcraciunas.screens.common.model
+
+data class PuzzleResult(
+    val id: Int?,
+    val rating: Int,
+    val success: Boolean,
+)
